Install a version specific console script
so that you will end up with a br# script, where the # is the version of Python referenced in the shebang. This way, if you install pip and pip3 install this package, you'll end up with a br script from whatever was pip installed last, and both a br2 and br3 that will run the specific version of python. I bet there's a better way, but this works for me for now, so there you go.