Changes so it works via tox
Don't force to python3 anymore, since tox is in control, and automatically handles running it under both python2 and python3. I had to make a few tweaks to the directory locations (relatively speaking), since tox runs the script from the root of the repo, and not the actual tests/ directory. I should probably use TESTDIR everywhere, instead of the relative path just to be consistent. Maybe later...