Point the tox workdir at a tmp location
Because I don't want it creating all the test environments inside .tox of my shared dropbox folder. NOTE: When I added custom environments, it didn't honor that, so I'll have to figure that out later on... Also, not sure why I no longer needed to pip install kenl380.pylib. I suppose it's doing that because of the setup.py dependency.